[DEV-4096] Move TAS matview as well
