+++ usaspending_api/common/matview_manager.py
@@ -1,7 +1,7 @@
 from collections import OrderedDict
 from django.conf import settings
 
-from usaspending_api.accounts.models import TASAutocompleteMatview
+from usaspending_api.search.models import TASAutocompleteMatview
 
 import usaspending_api.search.models as mv
